remove file only if that was a fifo

+++ Utils/Pipe/Resource/File.php
@@ -16,6 +16,8 @@ class File extends Stream implements ParameterInterface
 {
     use ParameterTrait;
 
+    protected $resourceIsFifo = false;
+
     public function openResourceHandle()
     {
         if (empty($this->resource)) {
@@ -23,6 +25,8 @@ class File extends Stream implements ParameterInterface
             if (!posix_mkfifo($this->resource, 0777)) {
                 throw new IOException();
             }
+
+            $this->resourceIsFifo = true;
         }
 
         if (!empty($this->parametersToParse)) {
@@ -44,6 +48,10 @@ class File extends Stream implements ParameterInterface
 
     public function __destruct()
     {
+        if (!$this->resourceIsFifo) {
+            return;
+        }
+
         file_exists($this->resource) && unlink($this->resource);
     }
 }
